agent/requirements.txt
fastapi>=0.115
python-dotenv>=1.0
uvicorn[standard]>=0.32
langchain-core>=0.3
pydantic>=2.0
langgraph>=0.2

# LLM providers — install all; each is only imported when its key is present
langchain-ollama>=0.2
langchain-openai>=0.2
langchain-anthropic>=0.3
langchain-google-genai>=2.0
langchain-groq>=0.2

# Tracing for the planner (fail-open at runtime if missing or misconfigured)
langfuse>=2.0

# Phase 3: hybrid retrieval
# psycopg2-binary: direct Postgres access for pgvector queries (Prisma doesn't support vector columns)
# openai: standalone SDK for embedding generation in the offline sync command
psycopg2-binary>=2.9
openai>=1.0

